// src/api/configurationValueType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Tubester API
 * To access protected endpoints, first log in:

[🔐 read only Login with Google](/api/auth/login/google?returnUrl=/swagger/index.html)

[🔐 write Login with Google](/api/auth/login/google/write?returnUrl=/swagger/index.html)
 * OpenAPI spec version: v1
 */

export type ConfigurationValueType = (typeof ConfigurationValueType)[keyof typeof ConfigurationValueType];

export const ConfigurationValueType = {
  String: 'String',
  Boolean: 'Boolean',
  Integer: 'Integer',
  Decimal: 'Decimal',
  Json: 'Json',
} as const;