// src/api/updateApplicationConfigurationRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Tubester API
 * To access protected endpoints, first log in:

[🔐 read only Login with Google](/api/auth/login/google?returnUrl=/swagger/index.html)

[🔐 write Login with Google](/api/auth/login/google/write?returnUrl=/swagger/index.html)
 * OpenAPI spec version: v1
 */
import type { ConfigurationValueType } from './configurationValueType';

export interface UpdateApplicationConfigurationRequest {
  value?: string;
  valueType?: ConfigurationValueType;
  /** @nullable */
  description?: string | null;
}